refactor: common cpc fn
