+++ 2020_RTSS_cpc_model_based_algorithm/src/prioritization_cpc_model.rs
@@ -97,161 +97,14 @@ fn assign_priority_to_cpc_model_core(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use lib::tests_helper::create_node;
-
-    fn create_sample_dag_not_consolidated() -> Graph<NodeData, i32> {
-        let mut dag = Graph::<NodeData, i32>::new();
-
-        //cX is the Xth critical node.
-        let c0 = dag.add_node(create_node(0, "execution_time", 10));
-        let c1 = dag.add_node(create_node(1, "execution_time", 10));
-        let c2 = dag.add_node(create_node(2, "execution_time", 10));
-
-        //create non-critical node.
-        //No distinction is made because of the complexity.
-        let n3 = dag.add_node(create_node(3, "execution_time", 3));
-        let n4 = dag.add_node(create_node(4, "execution_time", 2));
-        let n5 = dag.add_node(create_node(5, "execution_time", 3));
-        let n6 = dag.add_node(create_node(6, "execution_time", 1));
-        let n7 = dag.add_node(create_node(7, "execution_time", 1));
-        let n8 = dag.add_node(create_node(8, "execution_time", 3));
-
-        //create critical path edges
-        dag.add_edge(c0, c1, 1);
-        dag.add_edge(c1, c2, 1);
-
-        //create non-critical path edges
-        dag.add_edge(c0, n3, 1);
-        dag.add_edge(n3, c2, 1);
-        dag.add_edge(c0, n4, 1);
-        dag.add_edge(n4, n6, 1);
-        dag.add_edge(n5, n6, 1);
-        dag.add_edge(c0, n5, 1);
-        dag.add_edge(n5, n7, 1);
-        dag.add_edge(n6, n8, 1);
-        dag.add_edge(n7, n8, 1);
-        dag.add_edge(n8, c2, 1);
-
-        dag
-    }
-
-    fn create_sample_dag_complex() -> Graph<NodeData, i32> {
-        let mut dag = Graph::<NodeData, i32>::new();
-
-        //cX is the Xth critical node.
-        let c0 = dag.add_node(create_node(0, "execution_time", 10));
-        let c1 = dag.add_node(create_node(1, "execution_time", 10));
-        let c2 = dag.add_node(create_node(2, "execution_time", 10));
-
-        //create non-critical node.
-        //No distinction is made because of the complexity.
-        let n3 = dag.add_node(create_node(3, "execution_time", 2));
-        let n4 = dag.add_node(create_node(4, "execution_time", 2));
-        let n5 = dag.add_node(create_node(5, "execution_time", 3));
-        let n6 = dag.add_node(create_node(6, "execution_time", 1));
-        let n7 = dag.add_node(create_node(7, "execution_time", 1));
-        let n8 = dag.add_node(create_node(8, "execution_time", 3));
-
-        //create critical path edges
-        dag.add_edge(c0, c1, 1);
-        dag.add_edge(c1, c2, 1);
-
-        //create non-critical path edges
-        dag.add_edge(c0, n3, 1);
-        dag.add_edge(n3, c2, 1);
-        dag.add_edge(c0, n4, 1);
-        dag.add_edge(n4, n6, 1);
-        dag.add_edge(c0, n5, 1);
-        dag.add_edge(n5, n6, 1);
-        dag.add_edge(n5, n7, 1);
-        dag.add_edge(n6, n8, 1);
-        dag.add_edge(n7, n8, 1);
-        dag.add_edge(n8, c2, 1);
-
-        dag
-    }
-
-    ///DAG in Figure 2 (b) of the paper
-    fn create_sample_dag() -> Graph<NodeData, i32> {
-        let mut dag = Graph::<NodeData, i32>::new();
-        //cX is the Xth critical node.
-        let c0 = dag.add_node(create_node(0, "execution_time", 4));
-        let c1 = dag.add_node(create_node(1, "execution_time", 4));
-        let c2 = dag.add_node(create_node(2, "execution_time", 4));
-        let c3 = dag.add_node(create_node(3, "execution_time", 4));
-        let c4 = dag.add_node(create_node(4, "execution_time", 4));
-
-        //nY_X is the Yth preceding node of cX.
-        let n0_2 = dag.add_node(create_node(5, "execution_time", 2));
-        let n1_2 = dag.add_node(create_node(6, "execution_time", 1));
-        let n0_3 = dag.add_node(create_node(7, "execution_time", 3));
-        let n1_3 = dag.add_node(create_node(8, "execution_time", 2));
-        let n2_3 = dag.add_node(create_node(9, "execution_time", 1));
-        let n0_4 = dag.add_node(create_node(10, "execution_time", 3));
-        let n1_4 = dag.add_node(create_node(11, "execution_time", 2));
-        let n2_4 = dag.add_node(create_node(12, "execution_time", 2));
-
-        //create critical path edges
-        dag.add_edge(c0, c1, 1);
-        dag.add_edge(c1, c2, 1);
-        dag.add_edge(c2, c3, 1);
-        dag.add_edge(c3, c4, 1);
-
-        //create non-critical path edges
-        dag.add_edge(c0, n0_2, 1);
-        dag.add_edge(n0_2, c2, 1);
-        dag.add_edge(c0, n1_2, 1);
-        dag.add_edge(n1_2, c2, 1);
-        dag.add_edge(c0, n0_3, 1);
-        dag.add_edge(n0_3, c3, 1);
-        dag.add_edge(c1, n1_3, 1);
-        dag.add_edge(n1_3, c3, 1);
-        dag.add_edge(c1, n2_3, 1);
-        dag.add_edge(n2_3, c3, 1);
-        dag.add_edge(n0_3, n0_4, 1);
-        dag.add_edge(n0_4, c4, 1);
-        dag.add_edge(n1_3, n1_4, 1);
-        dag.add_edge(n1_4, c4, 1);
-        dag.add_edge(n2_3, n2_4, 1);
-        dag.add_edge(n2_4, c4, 1);
-
-        dag
-    }
+    use lib::tests_helper::create_dag_for_cpc;
 
     #[test]
     fn test_assign_priority_cpc_model_normal() {
-        let mut dag = create_sample_dag();
-        let expected_value = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
-
+        let mut dag = create_dag_for_cpc();
         assign_priority_to_cpc_model(&mut dag);
-        for node_i in dag.node_indices() {
-            assert_eq!(
-                dag[node_i].params["priority"],
-                expected_value[dag[node_i].id as usize]
-            );
-        }
-    }
 
-    #[test]
-    fn test_assign_priority_cpc_model_normal_dag_not_consolidated() {
-        let mut dag = create_sample_dag_not_consolidated();
         let expected_value = [0, 1, 2, 8, 6, 3, 7, 4, 5];
-
-        assign_priority_to_cpc_model(&mut dag);
-        for node_i in dag.node_indices() {
-            assert_eq!(
-                dag[node_i].params["priority"],
-                expected_value[dag[node_i].id as usize]
-            );
-        }
-    }
-
-    #[test]
-    fn test_assign_priority_cpc_model_normal_recursion() {
-        let mut dag = create_sample_dag_complex();
-        let expected_value = [0, 1, 2, 8, 6, 3, 7, 4, 5];
-
-        assign_priority_to_cpc_model(&mut dag);
         for node_i in dag.node_indices() {
             assert_eq!(
                 dag[node_i].params["priority"],
